test: 更改单元测试项目名称
