+++ test/UnitTest/Mock/MockTcpResponse.cs
@@ -2,7 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 // Website: https://github.com/LongbowExtensions/
 
-namespace UnitTestModbus;
+namespace UnitTest;
 
 static class MockTcpResponse
 {
