+++ test/UnitTest/Rtu/RtuClientTest.cs
@@ -3,13 +3,30 @@
 // Website: https://github.com/LongbowExtensions/
 
 using Microsoft.Extensions.DependencyInjection;
-using System.Net;
 
-namespace UnitTestModbus;
+namespace UnitTest;
 
-[Collection("MockTcpModbus")]
-public class TcpClientTest
+public class RtuClientTest
 {
+    [Fact]
+    public async Task Connect_Failed()
+    {
+        var sc = new ServiceCollection();
+        sc.AddModbusFactory();
+
+        var provider = sc.BuildServiceProvider();
+        var factory = provider.GetRequiredService<IModbusFactory>();
+        await using var client = factory.GetOrCreateRtuMaster("test", op =>
+        {
+            op.PortName = "COM3";
+        });
+
+        // 连接 Master
+        var connected = await client.ConnectAsync();
+        Assert.False(connected);
+        Assert.NotNull(client.Exception);
+    }
+
     [Fact]
     public async Task Connect_Exception()
     {
@@ -18,7 +35,7 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 未连接 Master 直接读取
         var ex = await Assert.ThrowsAnyAsync<InvalidOperationException>(async () =>
@@ -36,22 +53,44 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test", op =>
+        var client = factory.GetOrCreateRtuMaster("test", op =>
         {
-            op.ConnectTimeout = 1000;
-            op.LocalEndPoint = new(IPAddress.Any, 0);
+            op.PortName = "COM1";
+            op.DiscardNull = false;
+            op.BaudRate = 9600;
+            op.DataBits = 8;
+            op.Parity = System.IO.Ports.Parity.None;
+            op.StopBits = System.IO.Ports.StopBits.One;
+            op.ReadTimeout = 3000;
+            op.WriteTimeout = 3000;
+            op.RtsEnable = false;
+            op.DtrEnable = false;
+            op.Handshake = System.IO.Ports.Handshake.None;
+            op.ReadBufferSize = 4096;
+            op.WriteBufferSize = 2048;
+            op.RtsEnable = false;
+            op.DtrEnable = false;
         });
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        var connected = await client.ConnectAsync();
+        Assert.True(connected);
+
         var response = await client.ReadCoilsAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
 
-        await using var client2 = factory.GetOrCreateTcpMaster();
+        response = await client.ReadCoilsAsync(0x01, 0, 5);
+        Assert.NotNull(response);
+        Assert.Equal(5, response.Length);
+
+        var client2 = factory.GetOrCreateRtuMaster();
         Assert.NotEqual(client, client2);
 
-        factory.RemoveTcpMaster("test");
+        var client3 = factory.GetOrCreateRtuMaster("test");
+        Assert.Equal(client, client3);
+
+        await using var client4 = factory.RemoveRtuMaster("test");
     }
 
     [Fact]
@@ -62,10 +101,10 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        await client.ConnectAsync();
         var response = await client.ReadInputsAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
@@ -79,10 +118,10 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        await client.ConnectAsync();
         var response = await client.ReadHoldingRegistersAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
@@ -96,10 +135,10 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        await client.ConnectAsync();
         var response = await client.ReadInputRegistersAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
@@ -113,14 +152,14 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        await client.ConnectAsync();
         var response = await client.WriteCoilAsync(0x01, 0, true);
         Assert.True(response);
 
-        response = await client.WriteCoilAsync(0x01, 0, false);
+        response = await client.WriteCoilAsync(0x01, 1, false);
         Assert.True(response);
     }
 
@@ -132,10 +171,10 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        await client.ConnectAsync();
         var response = await client.WriteMultipleCoilsAsync(0x01, 0, [true, true, true, true, true, true, true, true, false, true]);
         Assert.True(response);
     }
@@ -148,10 +187,10 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        await client.ConnectAsync();
         var response = await client.WriteRegisterAsync(0x01, 0, 12);
         Assert.True(response);
     }
@@ -164,55 +203,11 @@ public class TcpClientTest
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateTcpMaster("test");
+        await using var client = factory.GetOrCreateRtuMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync("127.0.0.1", 502);
+        await client.ConnectAsync();
         var response = await client.WriteMultipleRegistersAsync(0x01, 0, [12, 0, 23, 0, 46, 0, 01, 02, 04, 05]);
         Assert.True(response);
-    }
-
-    [Fact]
-    public async Task ThreadSafe_Ok()
-    {
-        var clientCount = 1;
-        var taskCount = 5;
-        var sc = new ServiceCollection();
-        sc.AddModbusFactory();
-
-        var provider = sc.BuildServiceProvider();
-        var factory = provider.GetRequiredService<IModbusFactory>();
-
-        var clients = new List<IModbusClient>();
-        for (var index = 0; index < clientCount; index++)
-        {
-            var client = factory.GetOrCreateTcpMaster();
-            await client.ConnectAsync("127.0.0.1", 502);
-
-            clients.Add(client);
-        }
-
-        var results = new List<ushort[]>();
-        var tasks = clients.SelectMany(c =>
-        {
-            var tasks = new List<Task>();
-            for (int i = 0; i < taskCount; i++)
-            {
-                var task = Task.Run(async () =>
-                {
-                    var v = await c.ReadHoldingRegistersAsync(1, 0, 10);
-                    results.Add(v);
-                });
-                tasks.Add(task);
-            }
-            return tasks;
-        }).ToList();
-
-        await Task.WhenAll(tasks);
-
-        Assert.Equal(clientCount * taskCount, results.Count);
-
-        var failed = results.Count(i => i.All(v => v == 0));
-        Assert.Equal(0, failed);
     }
 }
