+++ test/UnitTest/Rtu/RtuOverTcpClientTest.cs
@@ -4,107 +4,44 @@
 
 using Microsoft.Extensions.DependencyInjection;
 
-namespace UnitTestModbus;
+namespace UnitTest;
 
-public class RtuClientTest
+[Collection("MockRtuOverTcpModbus")]
+public class RtuOverTcpClientTest
 {
-    [Fact]
-    public async Task Connect_Failed()
-    {
-        var sc = new ServiceCollection();
-        sc.AddModbusFactory();
-
-        var provider = sc.BuildServiceProvider();
-        var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test", op =>
-        {
-            op.PortName = "COM3";
-        });
-
-        // 连接 Master
-        var connected = await client.ConnectAsync();
-        Assert.False(connected);
-        Assert.NotNull(client.Exception);
-    }
-
-    [Fact]
-    public async Task Connect_Exception()
-    {
-        var sc = new ServiceCollection();
-        sc.AddModbusFactory();
-
-        var provider = sc.BuildServiceProvider();
-        var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
-
-        // 未连接 Master 直接读取
-        var ex = await Assert.ThrowsAnyAsync<InvalidOperationException>(async () =>
-        {
-            await client.ReadCoilsAsync(0x01, 0, 10);
-        });
-        Assert.NotNull(ex);
-    }
-
     [Fact]
     public async Task ReadCoilsAsync_Ok()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        var client = factory.GetOrCreateRtuMaster("test", op =>
-        {
-            op.PortName = "COM1";
-            op.DiscardNull = false;
-            op.BaudRate = 9600;
-            op.DataBits = 8;
-            op.Parity = System.IO.Ports.Parity.None;
-            op.StopBits = System.IO.Ports.StopBits.One;
-            op.ReadTimeout = 3000;
-            op.WriteTimeout = 3000;
-            op.RtsEnable = false;
-            op.DtrEnable = false;
-            op.Handshake = System.IO.Ports.Handshake.None;
-            op.ReadBufferSize = 4096;
-            op.WriteBufferSize = 2048;
-            op.RtsEnable = false;
-            op.DtrEnable = false;
-        });
+        await using var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        var connected = await client.ConnectAsync();
+        var connected = await client.ConnectAsync("127.0.0.1", 501);
         Assert.True(connected);
 
         var response = await client.ReadCoilsAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
-
-        response = await client.ReadCoilsAsync(0x01, 0, 5);
-        Assert.NotNull(response);
-        Assert.Equal(5, response.Length);
-
-        var client2 = factory.GetOrCreateRtuMaster();
-        Assert.NotEqual(client, client2);
-
-        var client3 = factory.GetOrCreateRtuMaster("test");
-        Assert.Equal(client, client3);
-
-        await using var client4 = factory.RemoveRtuMaster("test");
     }
 
     [Fact]
     public async Task ReadInputsAsync_Ok()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
+        var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync();
+        await client.ConnectAsync("127.0.0.1", 501);
         var response = await client.ReadInputsAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
@@ -114,14 +51,15 @@ public class RtuClientTest
     public async Task ReadHoldingRegistersAsync_Ok()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
+        var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync();
+        await client.ConnectAsync("127.0.0.1", 501);
         var response = await client.ReadHoldingRegistersAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
@@ -131,14 +69,15 @@ public class RtuClientTest
     public async Task ReadInputRegistersAsync_Ok()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
+        var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync();
+        await client.ConnectAsync("127.0.0.1", 501);
         var response = await client.ReadInputRegistersAsync(0x01, 0, 10);
         Assert.NotNull(response);
         Assert.Equal(10, response.Length);
@@ -148,14 +87,15 @@ public class RtuClientTest
     public async Task WriteCoilAsync_Ok()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
+        var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync();
+        await client.ConnectAsync("127.0.0.1", 501);
         var response = await client.WriteCoilAsync(0x01, 0, true);
         Assert.True(response);
 
@@ -167,14 +107,15 @@ public class RtuClientTest
     public async Task WriteMultipleCoilsAsync_Ok()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
+        var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync();
+        await client.ConnectAsync("127.0.0.1", 501);
         var response = await client.WriteMultipleCoilsAsync(0x01, 0, [true, true, true, true, true, true, true, true, false, true]);
         Assert.True(response);
     }
@@ -183,14 +124,15 @@ public class RtuClientTest
     public async Task WriteRegisterAsync()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
+        var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync();
+        await client.ConnectAsync("127.0.0.1", 501);
         var response = await client.WriteRegisterAsync(0x01, 0, 12);
         Assert.True(response);
     }
@@ -199,14 +141,15 @@ public class RtuClientTest
     public async Task WriteMultipleRegistersAsync_Ok()
     {
         var sc = new ServiceCollection();
+        sc.AddTcpSocketFactory();
         sc.AddModbusFactory();
 
         var provider = sc.BuildServiceProvider();
         var factory = provider.GetRequiredService<IModbusFactory>();
-        await using var client = factory.GetOrCreateRtuMaster("test");
+        var client = factory.GetOrCreateRtuOverTcpMaster("test");
 
         // 连接 Master
-        await client.ConnectAsync();
+        await client.ConnectAsync("127.0.0.1", 501);
         var response = await client.WriteMultipleRegistersAsync(0x01, 0, [12, 0, 23, 0, 46, 0, 01, 02, 04, 05]);
         Assert.True(response);
     }
